% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_parameters_old_to_new.R
\name{convert_parameters_old_to_new}
\alias{convert_parameters_old_to_new}
\title{Converts parameters from old to new format}
\usage{
convert_parameters_old_to_new(par)
}
\arguments{
\item{par}{vector of length 12 with the parameters in the following order:
yc, zc, s1, Tu, E0, E1, A0, A1, Tf, Tc, Tb, slope}
}
\value{
vector of length 12, with the PhenoFlex parameters yc, zc, s1, Tu, theta_star, theta_c, tau, pie_c, Tf, Tc, Tb, slope
with given temperature data and model parameters.
}
\description{
Takes the parameters used in  \link[chillR]{phenologyFitter} and brings it to
the format used in \code{\link{evaluation_function_meigo_nonlinear}}.
}
\details{
Original parameters include E0, E1, A0, A1 and convert those to theta_star, theta_c, tau and pie_c.
Parameters theta_star and tau depend on approximated intermediate variable sigma. Pie_c is
approximated using theta_star, theta_c and tau. Since the conversion
is an numerical approximation, the final results differ slightly from the "true" value.
This can be seen when first using \code{\link{convert_parameters}} and then convert the 
parameter again to the original format using this function. The error should  be
neglectible. Approximation is done using the function \link[nleqslv]{nleqslv}.

The conversion follows the approach documented in Fishman et al (1987) and Egea et al. (2021).
For more details consult equations 5 - 8 in Egea et al (2021)
}
\examples{
\dontrun{
par_old <-   c(40, 190, 0.5, 25, 279, 286.1, 47.7, 28, 4, 36, 4, 1.60)
par_new <- convert_parameters(par)
par_old_again <- convert_parameters_old_to_new(par)
}
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
