% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_ess.R
\name{custom_essr}
\alias{custom_essr}
\title{Enhanced scatter search with fixed bug from MEIGOR package}
\usage{
custom_essr(problem, opts = list(maxeval = NULL, maxtime = NULL), ...)
}
\arguments{
\item{problem}{list with starting values, upper ound, lower bound, and additional test ranges}

\item{opts}{list with controls for the essR algorithm}

\item{...}{allows to give further arguments of the evaluation function specified in the problem list}
}
\value{
list of fitted model
}
\description{
This is a fixed version of the essr package of MEIGOR. It was prone to internal errors, but actually now I cannot remember where it was.
}
\author{
Jose E. Egea, Lars Caspersen
}
