% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_gen_rel_change_scenario.R
\name{custom_gen_rel_change_scenario}
\alias{custom_gen_rel_change_scenario}
\title{Generates relative climate change scenarios based on extracted CMIP6 data}
\usage{
custom_gen_rel_change_scenario(
  downloaded_list,
  variable = c("Tmin", "Tmax"),
  scenarios = c(2050, 2085),
  reference_period = c(1986:2014),
  future_window_width = 30
)
}
\arguments{
\item{downloaded_list}{list of data.frames, generated using the
extract_cmip6_data function. Elements are named after the shared socioeconomic
pathway ('SSP') and global climate model ('GCM')}

\item{variable, }{vector with characters, specifies the variables for which
the relative change scenario should be calculated. By default: c('Tmin', 'Tmax'),
for minimum and maximum temperature. Also allows 'Prec' for precipitation}

\item{scenarios}{numeric vector, states the future years, for which the climate
change scenarios should be generated. Usually set to c(2050, 2085).}

\item{reference_period}{numeric vector specifying the years to be used as the
reference period. Usually set to to c(1986:2014).}

\item{future_window_width}{numeric, sets the window width of the running mean
calculation for the mean temperatures of the years indicated by scenarios}
}
\value{
data.frame for the calculated relative change scenarios, all locations, SSPs, timepoints, GCMs combined
}
\description{
Takes the extracted CMIP6 data and returns climate change scenarios, which
can then be used to generate weather data. Adjusted so that it also handles
precipitation
}
\examples{
\dontrun{
download_cmip6_ecmwfr(scenario = 'ssp1_2_6', 
                      area = c(55, 5.5, 47, 15.1),
                      user = 'write user id here',
                      key = 'write key here',
                      model = 'AWI-CM-1-1-MR',
                      frequency = 'monthly', 
                        variable = c('Tmin', 'Tmax', 'Prec'),
                      year_start = 2015, 
                      year_end = 2100)
                      
download_baseline_cmip6_ecmwfr(
    area = c(55, 5.5, 47, 15.1),
    model = 'AWI-CM-1-1-MR',
     variable = c('Tmin', 'Tmax', 'Prec'),
    frequency = 'monthly')
   
station <- data.frame(
      station_name = c('Zaragoza', 'Klein-Altendorf', 'Sfax',
      'Cieza', 'Meknes', 'Santomera'),
      longitude = c(-0.88,  6.99, 10.75, -1.41, -5.54, -1.05),
      latitude = c(41.65, 50.61, 34.75, 38.24, 33.88, 38.06))
      
extracted <- extract_cmip6_data(stations = station,
                                area =  c(52, -7, 33, 8),
                                variable = c('Tmin', 'Tmax', 'Prec'))

custom_gen_rel_change_scenario(extracted,
                        variable = c('Tmin', 'Tmax', 'Prec'),
                        scenarios = c(2050, 2085),
                        reference_period = c(1986:2014),
                        future_window_width = 30)

}

}
\author{
Lars Caspersen
}
