% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_prediction_with_failure.R
\name{ensemble_prediction_with_failure}
\alias{ensemble_prediction_with_failure}
\title{Make weighted mean prediction from scratch}
\usage{
ensemble_prediction_with_failure(
  par_list,
  confidence,
  modelfn,
  temp,
  return_se = TRUE,
  n_fail = 5,
  max_weight = NULL,
  ...
)
}
\arguments{
\item{par_list}{list of the parameters entries contain one read of the read with the function
LarsChill::load_fitting_result(), each entry is a repetition of the same cultivar
parameters are extracted from the 'xbest' entry within the individual list members
I assumed ten model parameters with theta_star and Tc being fixed, but you can change that in the code after the line par <- x$xbest
order of entries is identical with order in confidence (see next input)}

\item{confidence}{gives the weights to the individual predictions
numeric vector of same length as the par_list
assumes that bigger is better}

\item{modelfn}{function that takes one set of parameters and one entry of the seasonlist and returns a bloom date}

\item{temp}{seasonlist containing hourly temperature data for the predictions
generated with chillR::genSeasonList description}

\item{return_se}{logical, decides if standard deviation of the predictions around the weighted mean should be returned as well}

\item{n_fail}{numeric, decides the cut-off number of failure predictions of the weighted mean members so that the weighted mean also returns a failure
in case n_fail = 5: if we have 4 or less failure predictions --> get ignored and the weighted mean is calculated based on remaining results
if we have 5 or more failure predictions --> weighted mean is failure as well}

\item{max_weight}{by default NULL, when number between 0 and 1 supplied, it expresses how much
weight an individual prediction can get. Can prevent that one prediction dominates all the
remaining ones because the confidence score may be inflated}

\item{...}{further inputs for the \code{modelfn} argument}
}
\value{
when \code{return_se = TRUE} --> list with weighted mean, sd and individual model predictions
when \code{return_se = FALSE} --> vector with weighted means
}
\description{
Take model parameters, confidence scores for each set of parameters and calculates score-weighted mean prediction.
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
\keyword{PhenoFlex}
