% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_window_translation.R
\name{est_phen_gaps}
\alias{est_phen_gaps}
\title{Mutliple linear shifts of bloom timewindow}
\usage{
est_phen_gaps(target_df, target_col, split_col)
}
\arguments{
\item{target_df}{data.frame with phenological observations. Should contain a column called location and species}

\item{target_col}{character, specifies for which column the shifts should be calculated}

\item{split_col}{character, allows split into several groups if there are several phenological stages present. specifies the column which differentiates between the phenological stages}
}
\value{
data.frame with species, location, upper and lower limit in day of the year
}
\description{
Allows to shift properties of bloom timewindow from one location to another. 
Minimum requirement is at least one shared species between the locations.
}
\author{
Lars Caspersen
}
\keyword{utility}
