% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_seasonal_forecast.R
\name{extract_seasonal_forecast}
\alias{extract_seasonal_forecast}
\title{Opens downloaded seasonal forecast from the C3S database}
\usage{
extract_seasonal_forecast(file, target_lat = NULL, target_lon = NULL)
}
\arguments{
\item{file}{character, is the name of the file that needs to be extracted.
In case the file is not in the working directory, the value needs to include either an
absolute or relative path to the file.}

\item{target_lat, }{numeric, latitude of the location you want to extract. Be default set to NULL. If both target_lat and target_lon are NULL, then all the pixel values are extracted.}

\item{target_lon}{numeric, longitude of the location you want to extract. Be default set to NULL. If both target_lat and target_lon are NULL, then all the pixel values are extracted.}
}
\value{
data.frame containing the columns: Year, Month, Day, Hour, temp, unit, model, latitude, longitude, target_lat, target_lon.
Year, Month, Day, Hour indicate the date for the extracted value.
temp: contains the extracted temperature values
unit: indicates in what unit the temperatue is supplied (usually Kelvin)
model: indicates the individual model that provided the output. Usually, the seasonal forecast contain an
ensemble of models making the predictions. The underlying model is that generated the observation is shared among models, they are just different instances, as the forecast involves randomness.
latitude: original latitude of the pixel
longtidue: original longitude of the pixel
target_lat: latitude for the target point, that we wanted to extract
target_lon: longitude of the taret point, that we wanted to extract
}
\description{
Opens a downloaded netcdf file (.nc) and brings it into a data.frame format.
In case target latitude and longitude are provided, it extracts the values of the closest
pixel. Otherwise it will extract all pixel values. The function was developed for
precipitation but may also be compatible with other variables.
}
\examples{
\dontrun{

#download a simple dataset for a one week forecast
#takes one or two minutes
download_seasonal_forecast(year = c('1996'),
month = '11', 
area = c(51, 6.5, 50, 7.5),
leadtime_hour = seq(0, 24*7, by = 6), 
start_download = TRUE)

fname <- 'season-forecast_dwd21_2m_temperature_1996_11_1_0-168_51-6.5-50-7.5.nc'

extract_seasonal_forecast(file = fname, 
target_lat = 50.7,
target_lon = 7.1)

}

}
\author{
Lars Caspersen
}
