% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_downloaded_ssp.R
\name{format_downloaded_ssp}
\alias{format_downloaded_ssp}
\title{Bring downloaded CMIP6 data to chillR format}
\usage{
format_downloaded_ssp(downloaded)
}
\arguments{
\item{downloaded}{output from the \link[LarsChill]{download_cmip6_epwshiftr} function}
}
\value{
data.frame with extracted data for the specified locations. In case of 'tasmin' and
'tasmax' the original unit K gets converted to degree Celsius. In case of 'pr'
the original unit kg m-2 s-1 gets converted to mm day-1.
}
\description{
This function is best used after havin downloaded CMIP6 data via \link[LarsChill]{download_cmip6_epwshiftr}
}
\details{
At first data is queried using the function \link[epwshiftr]{init_cmip6_index}. If the 
user wants at first to get a feeling how much data files are found in the search, it
may be a good idea to run the query at first using directly the function. Terms for the search include:

\itemize{
 \item{Metric: Which weather variable should be downloaded. Commonly used are 'tasmin' for
 daily minimum temperature, 'tasmax' for daily maximum temperature and 'pr' for precipitation.} 
 \item{Experiment: defines the root experiment identifier. Commonly used Tier 1 experiments include: c("ssp126", "ssp245", "ssp370", "ssp585")}
 \item{Frequency: defines the timestep of the model output. Default is 'mon' for monthly}
 \item{Resolution: size of the pixels of the model output. Default is "100 km"}
 \item{...: there are more possible search terms which can include 'source' (name of model) or 'variant' (identifier for the individual run of the model output). For more
 details please consult the documentation of \link[epwshiftr]{init_cmip6_index}. Additional search terms can be specified at the end of the arguments via the '...' argument} 
}
}
\author{
Lars Caspersen
}
\keyword{utility}
