% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_combined_temp_response_plot.R
\name{gen_combined_temp_response_plot}
\alias{gen_combined_temp_response_plot}
\title{Draws temperature response plot from PhenoFlex parameters for several cultivars}
\usage{
gen_combined_temp_response_plot(
  performance_df,
  weather_list,
  temps = seq(-5, 50, by = 0.1),
  legend.pos = "bottom",
  col_palette = c("#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2", "#D55E00",
    "#CC79A7", "#999999"),
  par_names = c("yc", "zc", "s1", "Tu", "theta_star", "theta_c", "tau", "pie_c", "Tf",
    "Tc", "Tb", "slope"),
  chill_months = c(11:12, 1:2),
  heat_months = 1:3
)
}
\arguments{
\item{performance_df}{data.frame with at least the columns 'cultivar' and columns with the parameters name}

\item{weather_list}{list with hourly weather observations, needed to create the histogram with the temperature
frequency}

\item{temps}{numeric, vector containing the temperatures for which the temperature responses should be calculated}

\item{legend.pos}{character, by default 'bottom'. Specifies where the legend goes. Can take the same values as in ggplot2}

\item{col_palette}{character vector, contains the hexcode of the colors used to draw the temperature responses for each cultivar}

\item{par_names}{character vector, specifies the names of the parameters, by default the 'new' PhenoFlex parameters
c('yc', 'zc', 's1', 'Tu', 'theta_star', 'theta_c', 'tau', 'pie_c','Tf', 'Tc', 'Tb', 'slope')}

\item{chill_months}{numeric vector, indicating for which months the frequency of observed temperature should be calculated for 
the chill temperature response. By default set to c(11:12,1:3)}

\item{heat_months}{numeric vector, indicating for which months the frequency of observed temperature should be calculated for 
the heat temperature response. By default set to c(1:5)}
}
\value{
ggplot of the modeled temperature response
}
\description{
Returns a plot showing the chilling and forcing reaction of the PhenoFlex
model parameters for certain temperatures.
}
\details{
Furthermore, it can highlight which temperatures where observed in the 
data set used to generate the model parameters. The function can handle
the common PhenoFlex parameters (which include E0, E1, A0 and A1) and the
"new" parameters (which include theta_star, theta_c, tau and pie_c instead of the
earlier mentioned parameters).
}
\author{
Lars Caspersen
}
\keyword{utility}
