% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cmip6.R
\name{get_scenarioMIP_data}
\alias{get_scenarioMIP_data}
\title{Download climate change data of CMIP6 and extract for specified location}
\usage{
get_scenarioMIP_data(
  coordinates,
  start_year,
  end_year,
  metric,
  experiment,
  frequency = "mon",
  resolution = "100 km",
  activity = "ScenarioMIP",
  source = NULL,
  keep_downloaded = FALSE,
  path_download = "temp_cmip6",
  ...
)
}
\arguments{
\item{coordinates}{data.frame with the columns 'Longitude', 'Latitude' and 'id'. The data.frame
should contain the locations for which the user wants the climate change data to
be extracted. The 'id' column is used when reporting back the extracted climate change data.}

\item{start_year}{numeric, marks the start of the time series to be downloaded. Usually is 2015}

\item{end_year}{numeric, marks the end of the time series to be downloaded. Usually is 2100}

\item{metric}{character / vector  of characters indicating which variables should be downloaded. 
Common are 'tasmax' for near-surface maximum temperature, 'tasmin' for near-surface minimum temperature and
'pr' for precipitation. For more details please check also the documenation of \link[epwshiftr]{init_cmip6_index}}

\item{experiment}{character / vector or characters defining the root experiment identifier. Commonly used Tier 1 experiments include: c("ssp126", "ssp245", "ssp370", "ssp585")}

\item{frequency}{defines the timestep of the model output. Default is 'mon' for monthly}

\item{resolution}{size of the pixels of the model output. Default is "100 km"}

\item{keep_downloaded}{booloean, by default FALSE. If true, the function will not 
delete the downloaded .nc files. This makes sense when the user may want to 
use the climate change data for other locations. Prior to downloading, the function
checks if there are already files downloaded and if so it skipts the download step.}

\item{path_download}{character, defining the path relative to working directory, 
in which the .nc files should be (temporarily) safed}

\item{...}{allows the user to define further search terms for the query such as 
'source' or 'variant'.}
}
\value{
data.frame with extracted data for the specified locations. In case of 'tasmin' and
'tasmax' the original unit K gets converted to degree Celsius. In case of 'pr'
the original unit kg m-2 s-1 gets converted to mm day-1.
}
\description{
This function allows the used to download climate changed of the ScenarioMIP
project of CMIP6. Data is downloaded from the "Metagrid Search Interface" provided by ESGF with the
help of the \code{\link{epwshiftr}} package.
}
\details{
At first data is queried using the function \link[epwshiftr]{init_cmip6_index}. If the 
user wants at first to get a feeling how much data files are found in the search, it
may be a good idea to run the query at first using directly the function. Terms for the search include:

\itemize{
 \item{Metric: Which weather variable should be downloaded. Commonly used are 'tasmin' for
 daily minimum temperature, 'tasmax' for daily maximum temperature and 'pr' for precipitation.} 
 \item{Experiment: defines the root experiment identifier. Commonly used Tier 1 experiments include: c("ssp126", "ssp245", "ssp370", "ssp585")}
 \item{Frequency: defines the timestep of the model output. Default is 'mon' for monthly}
 \item{Resolution: size of the pixels of the model output. Default is "100 km"}
 \item{...: there are more possible search terms which can include 'source' (name of model) or 'variant' (identifier for the individual run of the model output). For more
 details please consult the documentation of \link[epwshiftr]{init_cmip6_index}. Additional search terms can be specified at the end of the arguments via the '...' argument} 
}
}
\author{
Lars Caspersen
}
\keyword{utility}
