% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp_response_df.R
\name{get_temp_response_df}
\alias{get_temp_response_df}
\title{Return temperature response}
\usage{
get_temp_response_df(par, temp_values)
}
\arguments{
\item{par}{traditional model parameters of PhenoFlex in the order yc, zc, s1, Tu, E0, E1, A0, A1, Tf, Tc, Tb, slope}

\item{temp_values}{numeric, vector containing the temperatures for which the temperature responses should be calculated}
}
\value{
data.frame with the columns "Temperature", "Chill_response" and "Heat_response"
}
\description{
Applies constant temperatures to chill / heat model and returns temperature response.
}
\details{
The output of the function is used in the temperature response plots, but it can be
handy to create those manually, especially when comparing temperature responses of
several cultivars or different set of parameters of same cultivar.
}
\author{
Lars Caspersen
}
\keyword{utility}
