% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_fitting_result.R
\name{load_fitting_result}
\alias{load_fitting_result}
\title{Load fitted PhenoFlex model}
\usage{
load_fitting_result(path, prefix)
}
\arguments{
\item{path}{character, specifying the location (relative to the working directory)}

\item{prefix}{character, added in front of the standard file name}
}
\value{
list with the fitted objects. Elements are named after the files (without numbering and prefix)
}
\description{
Allows to read fitted PhenoFlex models, which were saved using \link[LarsChill]{save_fitting_list}
}
\details{
This function reads the saved PhenoFlex model to a list.
}
\author{
Lars Caspersen
}
\keyword{utility}
