% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mva_bias_correction_forecast.R
\name{mva_bias_correction_forecast}
\alias{mva_bias_correction_forecast}
\title{Perform mean and variance adjustment (MVA) bias correction}
\usage{
mva_bias_correction_forecast(observed, predicted, unit_observed = "C")
}
\arguments{
\item{observed}{data.frame, contains observed temperature. Is usually the same format as 
when downloaded via chillR package. Should contain columns: Year, Month, Day, and a column with temperature. Valid options are:
Tmean, Temp, or combination of Tmin and Tmax.}

\item{predicted}{data.frame containing the forecast data. Usually the same format as when running the
extract_seaonal_forecast() function. Should contain columns: Year, Month, Day, Temp, unit}

\item{unit_observed, }{character, by default "C". Indicates the unit of the input temperature. Bias correction is done in Kelvin.
To indicate Fahrenheit, use "F". To indicate Kelvin use "K"}
}
\value{
data.frame, same format as "predicted", with additional column indicating correction factor and corrected (sub) daily temperature
}
\description{
Calculates mean biases and standard deviation for forecast data and local observed temperature
data. Performs bias correction on monthly means, then uses the ratio of corrected and uncorrected means
to scale the daily and sub-daily observations. Follows procedure outlined by: Manazanas et al. (2019) and
Leung et al. (1999). Returns a data.frame with corrected and original forecast data.
}
\details{
Bias correction does three steps for each month of the year:
\item{Calculate monthly means and sd for observed and forecasted data}
\item{Correct forecasted monthly mean with the formula: (mean_fcst_m_t - mean_fcst)*(sd_obs / sd_fcst) + mean_obs}
\item{Scale (sub) daily observation using share of corrected and uncorrected forecast mean}

mean_fcst_m_t = monthly mean of forecast member m, at year t
mean_fcst = monthly mean of all forecast members at all years
sd_obs = standard deviation of observed monthly means across all years
df_fcst = standard deviation of monthly means across all members m, across all years
mean_obs = monthly mean of observed temperature, across all years

Important note: as the function corrects using a factor, it is sensitive to sub-zero values. The function 
carries the mean correction out in Kelvin behin the scenes.
}
\examples{
\dontrun{

#download a simple dataset for a one week forecast
#takes one or two minutes

req <- download_seasonal_forecast(year = 1993:2000,
month = '11', 
area = c(51, 6.5, 50, 7.5),
leadtime_hour = seq(0, 24*30, by = 6), 
start_download = FALSE)

#run download
req <- download_seasonal_forecast(year = 1993:2000,
month = '11', 
area = c(51, 6.5, 50, 7.5),
leadtime_hour = seq(0, 24*30, by = 6), 
start_download = FALSE,
request_env = req)

#extract values
extract_seasonal_forecast(file = 'season-forecast_dwd21_2m_temperature_1993-2000_11_1_0-720_51-6.5-50-7.5.nc', 
target_lat = 50.7,
target_lon = 7.1)

#download local observation
long <- 7.0871843
lat <- 50.7341602
weather_dwd <- chillR::handle_dwd(action = 'list_stations', location = c(long, lat), 
                                  time_interval = c(19800101, 20251231))
                                  
data <- chillR::handle_dwd(action = "download_weather",
                           location = weather_dwd[1:3, "Station_ID"],
                           time_interval = c(19800101, 20241231),
                           stations_to_choose_from = 25,
                           station_list = weather_dwd,
                           drop_most = TRUE,
                           add.DATE = FALSE,
                            quiet = TRUE,
                            add_station_name = FALSE)
                            
data_clean <- chillR::handle_dwd(data)

observed <- data_clean$`Köln/Bonn`

predicted_bias_corrected <- mva_bias_correction_forecast(observed, predicted)

}

}
\author{
Lars Caspersen
}
