% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_ensemble_prediction.R
\name{pheno_ensemble_prediction}
\alias{pheno_ensemble_prediction}
\title{Get weighted average prediction of several PhenoFlex models}
\usage{
pheno_ensemble_prediction(
  par_list,
  confidence,
  modelfn,
  temp,
  return_se = TRUE,
  ...
)
}
\arguments{
\item{par_list}{the output of the phenology fitting done using MEIGO package}

\item{confidence}{numeric vector, containing the weights assign to the predictions. Assumes that higher value is more confidence. Values
get re-scaled by dividing by the sum of individual confidence scores.}

\item{modelfn}{function that takes one set of parameters and one entry of temp as input and returns bloomdate}

\item{temp}{SeasonList for the prediction, contains the hourly weather data of the weather stations}

\item{return_se}{boolean, by default set TRUE. Decides if sd of prediction and individual ensemble members predictions are returned as well}

\item{...}{further inputs for modelfn}
}
\value{
By default returns a list with three components: predicted, sd and individual_pred. Predicted
contains the weighted predictions of the ensemble members. sd contains the unweighted 
standard deviation of the members predictions. Individual_pred contains the ensemble members individual predictions
}
\description{
Outdated
}
\details{
This function takes the fitted models and temperature data and returns a weighted average prediction.
Weights are assigned using the confidence argument. 
The function assumes that theta_star and Tc are fixed. The
}
\author{
Lars Caspersen
}
\keyword{PhenoFlex}
