% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoflex_parnames_new.R
\docType{data}
\name{phenoflex_parnames_new}
\alias{phenoflex_parnames_new}
\title{PhenoFlex Parameter Names (New Version)}
\format{
A character vector with parameter names.
}
\usage{
data(phenoflex_parnames_new)
}
\description{
A character vector containing parameter names for the new version of Phenoflex.
}
\examples{
data(phenoflex_parnames_new)
head(phenoflex_parnames_new)
}
\keyword{datasets}
