% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenoflex_parnames_old.R
\docType{data}
\name{phenoflex_parnames_old}
\alias{phenoflex_parnames_old}
\title{PhenoFlex Parameter Names (Old Version)}
\format{
A character vector with parameter names.
}
\usage{
data(phenoflex_parnames_old)
}
\description{
A character vector containing parameter names for the old version of Phenoflex.
}
\examples{
data(phenoflex_parnames_old)
head(phenoflex_parnames_old)
}
\keyword{datasets}
