% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_predicted_days.R
\name{return_predicted_days}
\alias{return_predicted_days}
\title{Returns predicted bloomdays for a list of temperature time series and model parameters}
\usage{
return_predicted_days(par, modelfn, SeasonList, na_penalty = 365, ...)
}
\arguments{
\item{par}{traditional model parameters of PhenoFlex in the order yc, zc, s1, Tu, E0, E1, A0, A1, Tf, Tc, Tb, slope.}

\item{modelfn}{function used within the evaluation function to calculate the actual bloomday, often we use
the 'custom_GDH_wrapper' function for that}

\item{SeasonList}{list of hourly temperatures for the individual phenological seasons. Each element should contain a data.frame
with the columns "Temp" (for the hourly temperature) and "JDay" for the corresponding Julian day. Is usually
generated using \link[chillR]{genSeasonList}}

\item{na_penalty}{numeric, value which is used when the model fails to generate a prediction
for the bloom date. By default 365}

\item{...}{further arguments for modelfn}
}
\value{
numeric vector with the predicted bloom dates
}
\description{
This is a convenience function which allows to get the predicted bloom dates for a
set of model parameters and temperature observations. #'
}
\author{
Lars Caspersen
}
\keyword{utility}
