% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_fitting_list.R
\name{save_fitting_list}
\alias{save_fitting_list}
\title{Save fitted PhenoFlex model}
\usage{
save_fitting_list(fit_list, path, prefix)
}
\arguments{
\item{fit_list}{list with the fitted PhenoFlex models. Assumes that it is a list with at least one model. Elements should be
named, because these will be used when creating the file names.}

\item{path}{character, specifying the location (relative to the working directory)}

\item{prefix}{character, added in front of the standard file name}
}
\value{
list with the fitted objects. Elements are named after the files (without numbering and prefix)
}
\description{
Allows to save fitted PhenoFlex models. They can be loaded again using \link[LarsChill]{load_fitting_result}
}
\details{
This function saves the fitted model with all the additional information provided by MEIGOR function and
saves it to .txt files.
}
\author{
Lars Caspersen
}
\keyword{utility}
