% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/modified_comprehensive_precipitation_generator.R,
%   R/temperature_generation_rmawgen_prec.R
\name{modified_ComprehensivePrecipitationGenerator}
\alias{modified_ComprehensivePrecipitationGenerator}
\alias{temperature_generation_rmawgen_prec}
\title{This is a modified version of the \link[RMAWGEN]{ComprehensivePrecipitationGenerator}}
\usage{
modified_ComprehensivePrecipitationGenerator(
  station = c("T0001", "T0010", "T0099"),
  prec_all,
  mean_climate_prec = NULL,
  year_max = 1990,
  year_min = 1961,
  leap = TRUE,
  nmonth = 12,
  cpf = NULL,
  verbose = TRUE,
  p = 1,
  type = "none",
  lag.max = NULL,
  ic = "AIC",
  activateVARselect = FALSE,
  exogen = NULL,
  exogen_sim = NULL,
  is_exogen_gaussian = FALSE,
  year_max_sim = year_max,
  year_min_sim = year_min,
  mean_climate_prec_sim = NULL,
  onlygeneration = FALSE,
  varmodel = NULL,
  type_quantile = 3,
  qnull = NULL,
  valmin = 0.5,
  step = 0,
  n_GPCA_iteration = 0,
  n_GPCA_iteration_residuals = n_GPCA_iteration,
  sample = NULL,
  extremes = TRUE,
  exogen_all = NULL,
  exogen_all_col = station,
  no_spline = FALSE,
  nscenario = 1,
  seed = NULL,
  noise = NULL
)

temperature_generation_rmawgen_prec(
  weather,
  years,
  sim_years,
  temperature_scenario = data.frame(Tmin = rep(0, 12), Tmax = rep(0, 12)),
  seed = 99,
  check_temperature_scenario_type = TRUE,
  temperature_check_args = NULL,
  max_reference_year_difference = 5,
  warn_me = TRUE,
  remove_NA_scenarios = TRUE
)
}
\arguments{
\item{station}{station names}

\item{prec_all}{matrix with precipitation in mm for all stations. one column per station}

\item{mean_climate_prec}{by default NULL}

\item{year_max}{numeric, starting year}

\item{year_min}{numeric, end year}

\item{leap}{decides if leap years are included in simulated data}

\item{nmonth}{how many months are to be expected in input data}

\item{cpf}{don't know}

\item{verbose}{if message are printed while running function}

\item{p}{don't know}

\item{type}{dont't know}

\item{lag.max}{lag considered when simulating temperature and or precipitation}

\item{ic}{don't know}

\item{activateVARselect}{don't know}

\item{exogen}{if there are exogenous variables to be considered when generating weather}

\item{exogen_sim}{don't know}

\item{is_exogen_gaussian}{don't know}

\item{year_max_sim}{don't know}

\item{year_min_sim}{don't know}

\item{mean_climate_prec_sim}{don't know}

\item{onlygeneration}{too lazy to document}

\item{varmodel}{too lazy to document}

\item{type_quantile}{too lazy to document}

\item{qnull}{too lazy to document}

\item{valmin}{too lazy to document}

\item{step}{too lazy to document}

\item{n_GPCA_iteration}{too lazy to document}

\item{n_GPCA_iteration_residuals}{too lazy to document}

\item{sample}{too lazy to document}

\item{extremes}{too lazy to document}

\item{exogen_all}{too lazy to document}

\item{exogen_all_col}{too lazy to document}

\item{no_spline}{too lazy to document}

\item{nscenario}{too lazy to document}

\item{seed}{integer specifying the random seed for the weather generation.}

\item{noise}{too lazy to document}

\item{weather}{daily weather, as produced with the fix_weather
function. Can also be generated by other means, but should contain the columns
c("Month","Day","Year","Tmin","Tmax").}

\item{years}{vector of length 2 indicating the start and end year of the time
interval to be used for calibrating the temperature generator.}

\item{sim_years}{vector of length 2 indicating the start and end year of the time
interval for which temperatures are to be generated.}

\item{temperature_scenario}{can be one of three options:
1) a data.frame with two columns Tmin and Tmax and n_intervals (default: 12) rows containing
temperature changes for all time intervals, or absolute temperatures for these intervals.
2) a temperature scenario object, consisting of the following elements: 'data' = a data frame with
n_intervals elements containing the absolute or relative temperature information (as in input option 1);
'scenario_year' = the year the scenario is representative of; 'reference_year' = the year the scenario
is representative of; 'scenario_type' = the scenario type ('absolute' or 'relative' - if NA, this is
assigned automatically); 'labels' = and elements attached to the input temperature_scenario as an
element names 'labels'. A subset of these elements can also be specified, but 'data' must be present.
3) a (named or unnamed) list containing multiple objects of types 1 and 2. In this case, outputs are
generated for all scenarios.}

\item{check_temperature_scenario_type}{boolean variable specifying whether
temperature scenarios should be checked - and the scenario_type updated if necessary -
with the check_temperature_scenario function.}

\item{temperature_check_args}{list of arguments to be passed to the check_temperature_scenario function.
Check documentation of that function for details.}

\item{max_reference_year_difference}{for relative temperature scenarios, the maximum difference
between the reference years of the scenario and the weather record used for calibration (the median
of the two elements in the 'years' argument.}

\item{warn_me}{boolean variable specifying whether warnings should be shown. Defaults to TRUE.}

\item{remove_NA_scenarios}{boolean parameter indicating whether temperature scenarios that
contain NA values should be removed. Such scenarios would generate an error.}
}
\value{
list of data.frames containing the simulated weather, with columns c("YEARMODA",
"DATE","Year","Month","Day","Tmin","Tmax", "Prec"). If temperature_scenario is a list, the output
list contains simulated temperature records for all scenarios.

list of data.frames containing the simulated weather, with columns c("YEARMODA",
"DATE","Year","Month","Day","Tmin","Tmax", "Prec"). If temperature_scenario is a list, the output
list contains simulated temperature records for all scenarios.
}
\description{
The difference to the original version is, that the function crashes if the
spline interpolation of mean daily rainfall by month yields values close to zero.

Function to incorporate the temperature and precipitation generation function of the
RMAWGEN weather generator into chillR. The weather generator is calibrated
using the weather data.frame (years between years[1] and years[2]), and then
generates synthetic weather for a user-defined time frame (bounded by sim_years[1]
and sim_years[2]). Monthly change vectors for minimum and maximum temperatures and mean daily rainfall
(mm) can be specified to allow generation of temperature change scenarios.
}
\details{
This function is called by the wrapper function \code{\link{temperature_generation_rmawgen_prec}}. The rest of the 
documentation is copied from the original function.

Note that this function uses the temperature generation algorithms of the RMAWGEN
package. For more details, refer to the documentation of this package.
}
\author{
Lars Caspersen, Eike Luedeling

Lars Caspersen
}
\keyword{utility}
