% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean_with_fail.R
\name{weighted_mean_with_fail}
\alias{weighted_mean_with_fail}
\title{Process data.frame of predictions to weighted mean prediction}
\usage{
weighted_mean_with_fail(
  predicted,
  confidence,
  return_se = TRUE,
  n_fail = 5,
  max_weight = NULL,
  .progress = FALSE
)
}
\arguments{
\item{predicted}{data.frame with individual predictions, need to contain column called id
role of id column is to match predictions with the entries of the confidence input
so each cultivar would get their own id, can be the name, or something else
assume wide format --> so one column for each prediction
that means for one particular year we have ten columns with the ten predictions for that year (assuming you have ten repetitions)
function assumes that the names of the columns containing the predictions will be also present in the confidence input
so if the predictions are stored in columns R1 to R10, these column names need to be present in confidence, too.}

\item{confidence}{data.frame, needs to contain same columns as predicted (through predicted can have many more additional columns)
assume wide format, so 10 columns when ten repetitions, plus id
assumes that larger values mean more confidence
id entries need to match with id entries of the first input ('predicted')}

\item{return_se}{logical, decides if standard deviation of the predictions around the weighted mean should be returned as well}

\item{n_fail}{numeric, decides the cut-off number of failure predictions of the weighted mean members so that the weighted mean also returns a failure
in case n_fail = 5: if we have 4 or less failure predictions --> get ignored and the weighted mean is calculated based on remaining results
if we have 5 or more failure predictions --> weighted mean is failure as well}

\item{max_weight}{by default NULL, when number between 0 and 1 supplied, it expresses how much
weight an individual prediction can get. Can prevent that one prediction dominates all the
remaining ones because the confidence score may be inflated}

\item{.progress}{logical, if set TRUE than process par appears indicating the remaining time until the function completes}
}
\value{
same as 'predicted' but with additional column containing weighted mean and standard deviation
}
\description{
Helps to process already made predictions to weighted mean predictions. Has a routine how to handle predictions
of bloom failure. The function makes strong assumptions on the data frame of the predictions and the data frame
containing the weights (confidence)
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
