% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper_prepare_shift_plot.R
\name{wrapper_prepare_shift_plot}
\alias{wrapper_prepare_shift_plot}
\title{Prepare a the data.frame to make a shift plot}
\usage{
wrapper_prepare_shift_plot(
  shift_df,
  group_col,
  cut_share = c(0, 0.2, 0.4, 0.6, 0.8, 1)
)
}
\arguments{
\item{shift_df}{data.frame containing the summarized climate change impact projections and the baseline values. Assumes that it is already summarized (e.g. by median).
Assumes that the difference of future and baseline is summarized in the column "shift". Assumes baseline is summarized in column "baseline".}

\item{group_col}{character vector, containing the column names to group the shift data. Usually this includes the climate scenario,the scenario year and cultivar name.
All columns need to be present in shift_df.}

\item{cut_share}{numeric vector, segments of the shift plot. Should be between 0 and 1. For each interval, the function will calculate the percentage of shifts within or exceeding the interval#'}
}
\value{
data.frame, with additional columns for the intervals and the agreement among the climate projections regarding the intensity of the shift
}
\description{
Takes a data.frame as an input, usually the result of some kind of climate change impact projection.
Prepares data, so that I can make easily shift plots as in Caspersen et al. (in review), or the
analysis of apple phenology at lake Constance.
}
\details{
Makes relative strong assumptions on the structure of the input. Make sure to format it accordingly.
}
\author{
Lars Caspersen
}
\keyword{plotting}
