% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_essr.R
\name{custom_essr}
\alias{custom_essr}
\title{Optimization algorithm, taken from MEIGOR package}
\usage{
custom_essr(problem, opts = list(maxeval = NULL, maxtime = NULL), ...)
}
\arguments{
\item{problem}{list containing the initial parameter values x_0, the upper bound x_u, lower bound x_L and the 
evaluation function. You can also add inequality restraints to the problem statement}

\item{opts}{list of arguments controlling the optimization algorithm. 
Most important are maxeval, controlling the maximum number of evaluations before stopping the optimization, and
maxtime, maximum time in seconds for the optimizer to run}

\item{...}{Further arguments that get supplied to the evaluation function in the problem statement}
}
\value{
returns a list containing optimized parameters, intermediate solutions, final performance score and 
much more.
}
\description{
Optimization algorithm of enhanced scatter search. See the MEIGOR package for more details. Took the function
from the package because it seems that the MEIGOR package was deprecated and it was also painful to install
the package.
}
\details{
Check the MEIGOR documentation for more details
}
\author{
Lars Caspersen, \email{lars.caspersen@uni-bonn.de}
}
