% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_cmip6_ecmwfr.R
\name{download_cmip6_ecmwfr}
\alias{download_cmip6_ecmwfr}
\title{Download CMIP6 Data via the ecwfr package}
\usage{
download_cmip6_ecmwfr(
  scenario,
  n_try = 20,
  wait = 5,
  temporal_resolution = "monthly",
  variable = c("Tmin", "Tmax"),
  year_start = 2015,
  year_end = 2100,
  area = c(71, -15, 35, 50),
  month = 1:12,
  update_everything = FALSE,
  download_path = "cmip6_downloaded"
)
}
\arguments{
\item{scenario}{character, specifying the shared socioeconomic pathways (SSP) 
to be downloaded. Currently the values 'ssp1_2_6', 'ssp2_4_5', 'ssp3_7_0' and 
'ssp5_8_5' are the only legit options, which are also the standard scenarios 
of CMIP6}

\item{n_try}{numneric, number of repeated calls for the API. For more information see 'Details'}

\item{wait}{numeric, waiting in time between repeated function calls in seconds. For more
information see 'Details'}

\item{temporal_resolution}{character, can be either 'daily' or 'monthly'. Sets
if the downloaded CMIP6 data is in daily or monthly format}

\item{variable}{vector of characters, decides which variables get downloaded. 
Currently, the options "Tmin" (Daily minimum temperature in degree centigrade), 
"Tmax" (Daily maximum temperature in degree centigrade) 
and "Prec" (Daily sum of precipitation in mm) are the only valid options}

\item{year_start}{numeric, earliest year for downloaded CMIP6 data. By default set
to 2015}

\item{year_end}{numeric, latest year for downloaded CMIP6 data. By default set
to 2100}

\item{area}{numeric vector of length 4. Sets the spatial boundaries of the downloaded data.
Coordinates are supplied in the format:  c(maximum latitude, minimum longitude, minimum latitude, maximum longitude) which corresponds to the
northern extend, wester extend, southern extend and eastern extend of the area of interest}

\item{month}{numeric vector, sets which months should be downloaded. By default set to 1:12}

\item{update_everything}{logical, by default set to FALSE. When set FALSE already downloaded
scenarios with matching names are skipped. If set TRUE, then files are downloaded regardless if 
a file with the same name is already present}

\item{download_path}{character, sets the path for the download of the CMIP6 file. If not already present, then a 
new folder will be created. Path is relative to working directory}
}
\value{
NULL, the downloaded files are saved in the stated directory
}
\description{
Accesses the CMIP6 data of the Coperincus API via the ecwfr package. Saves the downloaded files as
.zip objects in the specified path.
}
\details{
Sometimes the server is not responding in time and the download fails because of that.
To avoid that the same line of code needs to be executed several times until
all the needed scenarios are downloaded, the call will be repeated in these cases. The argument n_calls sets the
maximum number of repeated calls. The argument wait sets the waiting time between repeated function calls.
}
\examples{
\dontrun{
}
}
\author{
Lars Caspersen, Antonio Picornell
}
