% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_temp_response_plot.R
\name{get_temp_response_plot}
\alias{get_temp_response_plot}
\title{Draws temperature response plot from PhenoFlex parameters}
\usage{
get_temp_response_plot(
  par,
  temp_values,
  hourtemps = NULL,
  chill_months = c(11:12, 1:2),
  heat_months = 1:5,
  weather_freq_plot = "histogram"
)
}
\arguments{
\item{par}{traditional model parameters of PhenoFlex in the order yc, zc, s1, Tu, E0, E1, A0, A1, Tf, Tc, Tb, slope}

\item{temp_values}{numeric, vector containing the temperatures for which the temperature responses should be calculated}

\item{hourtemps}{data.frame containing hourly temperature, by default set to NULL. If supplied, the columns
'Month' and 'Temp' should be present.}

\item{chill_months}{numeric vector, indicating for which months the frequency of observed temperature should be calculated for 
the chill temperature response. By default set to c(11:12,1:3)}

\item{heat_months}{numeric vector, indicating for which months the frequency of observed temperature should be calculated for 
the heat temperature response. By default set to c(1:5)}

\item{weather_freq_plot}{character, only applicable when `hourtemps` is supplied. 
Decides how temperature observations should be represented in the temperature
response plot. By default = 'histogram', which includes a histogram for the hourly
temperature. Other option is 'gradient' which represents the frequency of observed temperature
intervals in form of color gradient in the temperature response plot. Darker
color indicates most frequent temperature intervals, bright color less frequent ones.}
}
\value{
ggplot of the modeled temperature response
}
\description{
Returns a plot showing the chilling and forcing reaction of the PhenoFlex
model parameters for certain temperatures.
}
\details{
Furthermore, it can highlight which temperatures where observed in the 
data set used to generate the model parameters. The function can handle
the common PhenoFlex parameters (which include E0, E1, A0 and A1) and the
"new" parameters (which include theta_star, theta_c, tau and pie_c instead of the
earlier mentioned parameters).
}
\author{
Lars Caspersen
}
\keyword{plotting}
